Added currying request function, moved types to separate file, added biome.js for linting, created exception class and a constants file
