+++ package.json
@@ -15,6 +15,7 @@
     "define:change": "changeset",
     "release:apply": "changeset version",
     "release:publish": "changeset publish",
+    "lint": "npx @biomejs/biome check --write --unsafe ./src",
     "prebuild": "node -p \"'/*This file is auto generated during build, DO NOT CHANGE OR MODIFY */\\n\\nexport const ABACATE_PAY_VERSION = ' + JSON.stringify(require('./package.json').version) + ';'\" > src/version.ts",
     "build": "tsup ./src --out-dir ./dist --format cjs,esm --dts --silent --no-splitting --no-splitting --clean"
   },
@@ -30,6 +31,7 @@
     "node-fetch": "^3.3.2"
   },
   "devDependencies": {
+    "@biomejs/biome": "1.9.4",
     "@changesets/cli": "^2.27.10",
     "tsup": "^8.3.5",
     "typescript": "^5.7.2"
